requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
python-dateutil>=2.8.2
tkcalendar>=1.6.1
